refactor profile into components
