+++ resources/js/components/Profile/bottom/ProfileDownload.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="profile-card">
+        <p>Coming soon - Download My Data (easy)</p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'ProfileDownload'
+};
+</script>
+
+<style scoped>
+
+</style>
