+++ resources/js/components/Profile/bottom/ProfileTimeSeries.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="profile-card">
+        <p>Coming soon - Time-series chart of the users data</p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'ProfileTimeSeries'
+};
+</script>
+
+<style scoped>
+
+</style>
