+++ resources/js/components/Profile/middle/ProfileAwards.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="profile-card">
+        <p>Coming soon - Awards</p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'ProfileAwards'
+};
+</script>
+
+<style scoped>
+
+</style>
