+++ resources/js/components/Profile/middle/ProfileCategories.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="profile-card">
+        <p>Coming soon - Spider chart of the categories added</p>
+    </div>
+</template>
+
+<script>
+export default {
+    name: 'ProfileCategories'
+};
+</script>
+
+<style scoped>
+
+</style>
